adding logic of passing form as an array object
